% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anthro.01}
\alias{anthro.01}
\title{Raw data of a district level representative survey}
\format{
A tibble with 1191 rows and 11 columns.\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{area} \tab Location where the survey took place \cr
   \emph{dos} \tab Survey date \cr
   \emph{cluster} \tab Primary sampling unit \cr
   \emph{team} \tab Enumerator IDs \cr
   \emph{sex} \tab Sex, "m" = boys, "f" = girls \cr
   \emph{dob} \tab Date of birth \cr
   \emph{age} \tab Age in months, typically estimated using local event calendars \cr
   \emph{weight} \tab Weight (kg) \cr
   \emph{height} \tab Height (cm) \cr
   \emph{edema} \tab Edema, "n" = no, "y" = yes \cr
   \emph{muac} \tab Mid-upper arm circumference (mm) \cr
}
}
\usage{
anthro.01
}
\description{
#' \code{anthro.01} is about a two-stage and PPS cluster sampling survey data
conducted in two district following the SMART survey methodology in two
livelihood zones. The location information was anonymized for confidentiality.
}
\examples{
anthro.01

}
\keyword{datasets}
