% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anthro.02}
\alias{anthro.02}
\title{Province representative survey conducted in Mozambique}
\format{
A tibble with 2267 rows and 14 columns.\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{province} \tab The administrative unit (admin 1) where data was collected. \cr
   \emph{strata} \tab Rural and Urban \cr
   \emph{cluster} \tab Primary sampling unit \cr
   \emph{sex} \tab Sex, "m" = boys, "f" = girls \cr
   \emph{age} \tab calculated age in months with two decimal places \cr
   \emph{weight} \tab Weight (kg) \cr
   \emph{height} \tab Height (cm) \cr
   \emph{edema} \tab Edema, "n" = no, "y" = yes \cr
   \emph{muac} \tab Mid-upper arm circumference (mm) \cr
   \emph{wtfactor} \tab Survey weights \cr
   \emph{wfhz} \tab Weight-for-height z-scores with 3 decimal places \cr
   \emph{flag_wfhz} \tab Flagged observations. 1=flagged, 0=not flagged \cr
   \emph{mfaz} \tab MUAC-for-age z-scores with 3 decimal places \cr
   \emph{flag_mfaz} \tab Flagged observations. 1=flagged, 0=not flagged \cr
}
}
\usage{
anthro.02
}
\description{
\code{anthro.02} is about a household budget survey conducted in Mozambique in
2019/2020, known as IOF (\emph{Inquérito ao Orçamento Familiar} in portuguese).
The data is publicly available \href{https://mozdata.ine.gov.mz/index.php/catalog/88#metadata-data_access}{here}.
The survey had a module on nutrition with anthropometric measurements taken
from children age 0-59 months for weight-for-height and 6-59 months for MUAC.
\emph{IOF} is a cluster and PPS-based, survey, with sampling done in two stages,
designed to give representative estimates at province level. Its data
collection spans for a period of 12 months, with anthropometric measurements
taken during that period too. Read the \href{https://mozdata.ine.gov.mz/index.php/catalog/88#metadata-sampling}{Bureau of Statistic's website on IOF} for
more details.

\code{anthro.02} has been processed for this package's purpose.
}
\examples{
anthro.02


}
\keyword{datasets}
