% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anthro.03}
\alias{anthro.03}
\title{District level survey data with Standard Deviation classified as problematic}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 943 rows and 11 columns.
}
\usage{
anthro.03
}
\description{
District level survey data with Standard Deviation classified as problematic
}
\keyword{datasets}
