% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anthro.04}
\alias{anthro.04}
\title{MUAC data from a community-based sentinel site from an anonymized location}
\format{
A tibble of 3,002 x 8.\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{province} \tab  \cr
   \emph{cluster} \tab Primary sampling unit \cr
   \emph{sex} \tab Sex, "m" = boys, "f" = girls \cr
   \emph{age} \tab calculated age in months with two decimal places \cr
   \emph{muac} \tab Mid-upper arm circumference (mm) \cr
   \emph{edema} \tab Edema, "n" = no, "y" = yes \cr
   \emph{mfaz} \tab MUAC-for-age z-scores with 3 decimal places \cr
   \emph{flag_mfaz} \tab Flagged observations. 1=flagged, 0=not flagged \cr
}
}
\usage{
anthro.04
}
\description{
Data in \code{anthro.04} was generated from a community-based sentinel site of three provinces.
Each province data set holds different scenarios that informs the appropriate analysis
approach to follow. One province (province 3) has its MFAZ standard deviation and age
ratio tests classified as problematic. Another province (province 2) has its age ratio
classified as problematic, but with a within range standard deviation. Lastly, province 1
has both tests falling within range of nor problematic. The data is used to test the
performance of \verb{[compute_muac_prevalence()]} based when used on a multiple survey areas
data that may or not have on the aforementioned test that may then warrant a different
analysis approach, as the function is designed for.
}
\keyword{datasets}
