% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anthro_data}
\alias{anthro_data}
\title{A survey data with anthropometric measurements, including edema}
\format{
A data frame with 11 columns and 7650 rows\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{area} \tab Location where the survey took place \cr
   \emph{dos} \tab Survey date \cr
   \emph{cluster} \tab Primary sampling unit \cr
   \emph{team} \tab Enumerator IDs \cr
   \emph{sex} \tab Sex, "m" = boys, "f" = girls \cr
   \emph{dob} \tab Date of birth \cr
   \emph{age} \tab Age in months, typically estimated using local event calendars \cr
   \emph{weight} \tab Weight (kg) \cr
   \emph{height} \tab Height (cm) \cr
   \emph{edema} \tab Edema, "n" = no, "y" = yes \cr
   \emph{muac} \tab Mid-upper arm circumference (mm) \cr
}
}
\usage{
anthro_data
}
\description{
A SMART survey data with a data structure typical of data of ENA for SMART
Software. Location information anonymised.
}
\examples{

anthro_data

}
\keyword{datasets}
