% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_wfhz.R
\name{apply_probit_approach}
\alias{apply_probit_approach}
\title{Compute global, severe and moderate acute malnutrition prevalence using PROBIT approach.}
\usage{
apply_probit_approach(x, .status = c("gam", "sam"))
}
\arguments{
\item{x}{A double vector containing the z-score values}

\item{.status}{A choice on the nutritional status you wish to apply the PROBIT approach
on. Default is "gam" for global acute malnutrition.}
}
\value{
A numeric value (double) corresponding to the point prevalence estimate.
}
\description{
This approach is only applied for when WFHZ standard deviation's is problematic. The
PROBIT approach estimates the prevalence of acute malnutrition indirectly by computing
the area under the tail of the curve from negative infinitive to the given threshold
through the cumulative normal distribution function using the mean and standard deviation.
}
