% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_scorers.R
\name{assign_penalty_points_age_sex_ratio}
\alias{assign_penalty_points_age_sex_ratio}
\title{Assign a penalty point for the amount of selection biases in age and sex ratios}
\usage{
assign_penalty_points_age_sex_ratio(x)
}
\arguments{
\item{x}{A numeric vector containing p-values from either age or sex ratio
test results.}
}
\value{
A numeric vector with the corresponding penalty points (scores) according
to the classification.
}
\description{
The function assigns a penalty score for a age and sex ratio's test classification.
The score range varies between 0 (when "Excellent") to 10 (when "Problematic") for
both, according to the \href{https://smartmethodology.org/}{ENA for SMART software}.
}
\examples{

 ## A vector storing age ratio or sex ratio p-values' classification ----
 x <- c("Excellent", "Problematic", "Acceptable", "Good")
 ## Apply the function ----
 assign_penalty_points_age_sex_ratio(x)

}
