% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_scorers.R
\name{assign_penalty_points_flags_and_sd}
\alias{assign_penalty_points_flags_and_sd}
\title{Assign a penalty point for the amount of proportion flagged data and standard deviation}
\usage{
assign_penalty_points_flags_and_sd(x)
}
\arguments{
\item{x}{A character vector containing the test classifications of proportion
of flagged data and the value of standard deviation.}
}
\value{
A numeric vector with the corresponding penalty points (scores) according
to the classification.
}
\description{
The function assigns a penalty score for a given category of test classification.
The score range varies between 0 (when "Excellent") to 20 (when "Problematic") for
both flagged data and standard deviation. This was borrowed from the
\href{https://smartmethodology.org/}{ENA for SMART software}
In the SMART Methodology, flagged data and standard deviation are tho test
criteria that gets the highest penalty scores, so it is here.
}
\examples{

## Sample data ----
x <- c("Excellent", "Problematic", "Acceptable", "Good")
## Apply the function ----
assign_penalty_points_flags_and_sd(x)

}
