% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_scorers.R
\name{assign_penalty_points_skew_kurt}
\alias{assign_penalty_points_skew_kurt}
\title{Assign a penalty point for the amount of issues in Skweness and Kurtosis}
\usage{
assign_penalty_points_skew_kurt(x)
}
\arguments{
\item{x}{A numeric vector containing Skewness or Kurtosis test results classification.}
}
\value{
A numeric vector with the corresponding penalty points (scores) according
to the classification.
}
\description{
The function assigns a penalty score for a Skewness and Kurtosis test classification.
The score range varies between 0 (when "Excellent") to 5 (when "Problematic") for
both, according to the \href{https://smartmethodology.org/}{ENA for SMART software}.
}
\examples{

 ## A vector storing Skewness or Kurtosis test classification ----
 x <- c("Excellent", "Problematic", "Acceptable", "Good")
 ## Apply the function ----
 assign_penalty_points_skew_kurt(x)

}
