% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wasting_cases_muac}
\alias{wasting_cases_muac}
\alias{wasting_cases_whz}
\alias{wasting_cases_combined}
\title{Case-Definition: is an observation acutely malnourished?}
\usage{
wasting_cases_muac(muac, edema = NULL, cases = c("gam", "sam", "mam"))

wasting_cases_whz(zscore, edema = NULL, cases = c("gam", "sam", "mam"))

wasting_cases_combined(
  zscore,
  muac,
  edema = NULL,
  cases = c("cgam", "csam", "cmam")
)
}
\arguments{
\item{muac}{An integer vector containing MUAC measurements in mm.}

\item{edema}{A character vector of "y" = Yes, "n" = No bilateral edema.
Default is NULL.}

\item{cases}{A choice of wasting case definition you wish to apply. For combined
acute malnutrition with \code{\link[=wasting_cases_combined]{wasting_cases_combined()}} cases options are:
c("cgam", "csam", "cmam").}

\item{zscore}{A double vector containing weight-for-height zscores with 3
decimal places.}
}
\value{
A dichotomous vector of 1=Yes and 0=No.
}
\description{
\code{\link[=wasting_cases_muac]{wasting_cases_muac()}}, \code{\link[=wasting_cases_whz]{wasting_cases_whz()}} and \code{\link[=wasting_cases_combined]{wasting_cases_combined()}}
help you get through with your wasting case definition for each observation.
It should be used inside dplyr::mutate() or base::transform(). It was designed
to be used inside \code{\link[=define_wasting]{define_wasting()}}.
}
\examples{
# example code

}
