% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdc_muac_weighting.R
\name{cdc_apply_age_weighting}
\alias{cdc_apply_age_weighting}
\title{Correct the observed MUAC prevalence when there is an unbalanced sample
between children under 2 and over two years old}
\usage{
cdc_apply_age_weighting(muac, age, edema = NULL, status = c("sam", "mam"))
}
\arguments{
\item{muac}{An integer vector containing MUAC measurements in mm.}

\item{age}{A double vector containing age in months with at least 2 decimal
places.}

\item{edema}{Optional. If given, it should be a character vector of "y" = Yes,
"n" = No bilateral edema.}

\item{status}{If you wish to get the prevalence/proportions of severe or
moderate acute malnutrition. Set \verb{status = "sam" or status = "mam"} for the
former or latter, respectively.}
}
\description{
As documented in the SMART MUAC tool and in the literature, MUAC shows a known
bias towards younger children. In a balanced sample, it is expected to have
nearly two thirds of the sample to be of children over two years old. If too
few older children are included in the sample, the weighted tool should be used.

\code{cdc_apply_age_weighting()} does that. It takes the proportion of children
under 2 and adds to the product of 2 times the proportion of children over two,
then divided by 3. The use of this function is informed by the output of
\code{\link[=age_ratio_test]{age_ratio_test()}}. There is difference between this function and that in the
SMART plausibility check. Consider reading the documentation before use.
}
