% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdc_muac_weighting.R
\name{cdc_classify_wasting}
\alias{cdc_classify_wasting}
\title{A helper function to classify nutritional status into SAM, MAM or not wasted}
\usage{
cdc_classify_wasting(muac, edema = NULL)
}
\arguments{
\item{muac}{An integer vector containing MUAC values. They should be in
millimeters.}

\item{edema}{Optional. Its a vector containing data on bilateral pitting
edema coded as "y" for yes and "n" for no.}
}
\description{
\code{cdc_classify_wasting()} is used a helper inside \code{\link[=cdc_apply_age_weighting]{cdc_apply_age_weighting()}} to
classify nutritional status into "sam", "mam" or "not wasted" and then the
vector returned is used downstream to calculate the proportions of children
with severe and moderate acute malnutrition.
}
