% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_checkers.R
\name{check_plausibility_crude_muac}
\alias{check_plausibility_crude_muac}
\title{Plausibility check: Crude MUAC data.}
\usage{
check_plausibility_crude_muac(df, flags, sex, muac, area)
}
\arguments{
\item{df}{A data frame object returned by \code{\link[=process_muac_data]{process_muac_data()}}}

\item{flags}{A character vector telling whether or not an observation is an
outlier.}

\item{sex}{A vector telling whether a given child is a boy or girl.}

\item{muac}{A vector containing MUAC measurements. This is used by
\link[nipnTK:digitPreference]{nipnTK::digitPreference}.}

\item{area}{A vector with values on where was the data collected. If you are
analyzing a data set with just one area, provide it anyway to
\code{check_plausibility_crude_muac()}}
}
\description{
\code{check_plausibility_crude_muac()} lets you know the quality of your crude MUAC
data. This should used only when the variables age and or sex are missing from
your dataset, as not having them cannot allow you calculate MUAC-for-age
zscores as required in \code{\link[zscorer:addWGSR]{zscorer::addWGSR()}}.
}
\examples{

quality <- data.01 |>
process_muac_data(
sex = "sex",
muac = "muac",
.recode_sex = TRUE,
.recode_muac = FALSE,
unit = "none"
) |>
check_plausibility_crude_muac()

## View object ----
quality

}
