% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_checkers.R
\name{check_plausibility_mfaz}
\alias{check_plausibility_mfaz}
\title{Plausibility check: MUAC-for-age zscores}
\usage{
check_plausibility_mfaz(df, flags, sex, muac, age, area)
}
\arguments{
\item{df}{A data frame object returned by \code{\link[=process_muac_data]{process_muac_data()}}.}

\item{flags}{A character vector telling whether or not an observation is an
outlier.}

\item{sex}{A vector telling whether a given child is a boy or girl.}

\item{muac}{A vector containing MUAC measurements. This is used by
\link[nipnTK:digitPreference]{nipnTK::digitPreference}.}

\item{age}{A vector containing children age in months.}

\item{area}{A vector with values on where was the data collected. If you are
analyzing a data set with just one area, provide it anyway to \code{check_plausibility_mfaz()}}
}
\value{
A summarized data frame of 18 columns containing quality checks
statistics and respective classification. The classification is the same as that
of SMART Methodology.
}
\description{
\code{check_plausibility_mfaz()} lets you know the quality of your data, based on
the statistics around MFAZ.
}
\examples{

quality <- data.01 |>
process_age() |>
process_muac_data(
sex = "sex",
age = "age",
muac = "muac",
.recode_sex = TRUE,
.recode_muac = TRUE,
unit = "cm"
) |>
check_plausibility_mfaz(
flags = "flags",
sex = "sex",
muac = "muac",
age = "age",
area = area
)

quality

}
