% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_checkers.R
\name{check_plausibility_whz}
\alias{check_plausibility_whz}
\title{Plausibility check: Weight-for-Height z-scores}
\usage{
check_plausibility_whz(df, sex, age, wgt, hgt, flags, area)
}
\arguments{
\item{df}{A data frame object returned by \code{\link[=process_whz_data]{process_whz_data()}}.}

\item{sex}{A vector telling whether a given child is a boy or girl.}

\item{age}{A vector containing children age in months.}

\item{wgt}{A vector containing weight measurements in kilograms.}

\item{hgt}{A vector containing height measurements in centimeters.}

\item{flags}{A character vector telling whether or not an observation is an
outlier.}

\item{area}{A vector with values on where was the data collected. If you are
analyzing a data set with just one area, provide it anyway to \code{check_plausibility_whz()}}
}
\value{
A summarized data frame of 18 columns containing quality checks
statistics and respective classification. The classification is the same as that
of SMART Methodology.
}
\description{
\code{check_plausibility_whz()} lets you know the quality of your data, based on
statistics around WHZ, but without including MUAC, as it is on SMART plausibility
report.
}
\examples{
quality <- data.01 |>
process_age() |>
process_whz_data(
sex = "sex",
wgt = "weight",
hgt = "height",
.recode_sex = TRUE
) |>
check_plausibility_whz(
flags = "flags",
sex = "sex",
age = "age",
wgt = "weight",
hgt = "hgt",
area = area
)

quality


}
