% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_classifiers.R
\name{classify_overall_quality}
\alias{classify_overall_quality}
\title{Get the overall data quality classification}
\usage{
classify_overall_quality(df)
}
\arguments{
\item{df}{A data frame containing a vector with the quality scores generated by
\code{\link[=get_quality_score]{get_quality_score()}}.}
}
\value{
A character vector of the same length, but a different width as the
input \code{df} is returned with a new column called \code{quality_class}.
}
\description{
\code{classify_overall_quality()} helps you in knowing the overall status of your
data quality. It classifies the overall score generated by
\code{\link[=get_quality_score]{get_quality_score()}} into four categories, as it is done in the
\href{https://smartmethodology.org/}{SMART Methodology},
namely: "Excellent", "Good", "Acceptable" and "Problematic". Beware that
the overall classification should be used as an indication to further
scrutinize of data before taking the decision to validate or invalidate the
results.
}
\examples{
## A sample vector of quality score ----

df <- data.frame(
flagged_class = "excellent",
age_ratio_class = "Good",
sex_ratio_class = "Problematic",
dps_class = "Excellent",
sd_class = "Excellent",
skewness_class = "Good",
kurtosis_class = "Acceptable"
)

## Apply get_quality_score()
w <- get_quality_score(df, method = "mfaz")

## Apply classify_overall_quality()
classify_overall_quality(w)

### You can choose to chain the functions with a pipe operator ----
df |>
get_quality_score(method = "mfaz") |>
classify_overall_quality()

}
