% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_classifiers.R
\name{classify_sd}
\alias{classify_sd}
\title{Classify how much high is the value of standard deviation}
\usage{
classify_sd(sd, type = c("zscore", "crude"))
}
\arguments{
\item{sd}{A numeric vector containing values for standard deviation of the
method you wish the work on.}

\item{type}{The method to which you wish to classify how much high is the
value of standard deviation. A choice between "zscore" MFAZ or WHZ and
"crude" for crude MUAC.}
}
\value{
A character vector with the correspondent classification.
}
\description{
\code{classify_sd()} helps you to know the magnitude of the data's standard
deviation. You can use this function for either WHZ, MFAZ or crude MUAC.
Cut-offs for WHZ are based on the \href{https://smartmethodology.org/}{SMART Methodology}.
Cut-offs for MFAZ are also based on SMART, but informed by
\href{https://doi.org/10.1111/mcn.13478}{Bilukha, O., & Kianian, B. (2023).}.
For crude MUAC, the cut-offs are based on the
\href{https://www.ipcinfo.org/ipcinfo-website/resources/ipc-manual/en/}{IPC AMN guidelines}
}
\examples{

## Have a vector with standard deviation ----
sdvalues <- seq(0.7, 1.3, by = 0.001) |>
sample(size = 9, replace = TRUE)

## Apply the function with `type = "zscore` ----
classify_sd(sdvalues, type = "zscore")

## Using `type = "crude"` ----
### Create sample data ----
sdvalues <- seq(9, 30, by = 2) |>
sample(size = 20, replace = TRUE)

### Apply the function with `type = "crude"` ----
classify_sd(sdvalues, type = "crude")

}
