% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_classifiers.R
\name{classify_skew_kurt}
\alias{classify_skew_kurt}
\title{Classify how much high is the value of Skewness and Kurtosis}
\usage{
classify_skew_kurt(sk)
}
\arguments{
\item{sk}{A numeric vector containing values of either Skewness or Kurtosis.}
}
\value{
A character vector with the correspondent classification.
}
\description{
\code{classify_skew_kurt()} helps you to know the magnitude of the Skewness and
Kurtosis from your data. This is only useful for WHZ and MFAZ. The function
works on the results yielded by \code{\link[nipnTK:skewKurt]{nipnTK::skewKurt()}}.
Cut-offs for WHZ are based on the \href{https://smartmethodology.org/}{SMART Methodology}.
}
\examples{

#Have a numeric vector storing values for skewness or kurtosis ----
sk <- seq(-5, 1, by = 0.05) |> sample(size = 20, replace = TRUE)

# Apply function
classify_skew_kurt(sk)

}
