% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{classify_wasting}
\alias{classify_wasting}
\title{A helper function to classify nutritional status}
\usage{
classify_wasting(muac, edema = NULL)
}
\arguments{
\item{muac}{An integer vector containing MUAC values. They should be in
millimeters.}

\item{edema}{Optional. Its a vector containing data on billateral pitting
edema coded as "y" for yes and "n" for no.}
}
\description{
\code{classify_wasting()} is used a helper inside \code{\link[=apply_CDC_age_weighting]{apply_CDC_age_weighting()}} to
classify nutritional status into "sam", "mam" or "not wasted" and then the
vector returned is used downstream to calculate the proportions of children
with severe and moderate acute malnutrition.
}
