% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_muac.R
\name{classify_wasting_for_cdc_approach}
\alias{classify_wasting_for_cdc_approach}
\title{A helper function to classify nutritional status into SAM, MAM or not wasted}
\usage{
classify_wasting_for_cdc_approach(muac, .edema = NULL)
}
\arguments{
\item{muac}{An integer vector containing MUAC values. They should be in
millimeters.}

\item{.edema}{Optional. Its a vector containing data on bilateral pitting
edema coded as "y" for yes and "n" for no.}
}
\description{
\code{classify_wasting_for_cdc_approach()} is used a helper inside
\code{\link[=apply_cdc_age_weighting]{apply_cdc_age_weighting()}} to classify nutritional status into "sam", "mam"
or "not wasted" and then the vector returned is used downstream to calculate
the proportions of children with severe and moderate acute malnutrition.
}
