% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_combined.R
\name{compute_pps_based_combined_prevalence}
\alias{compute_pps_based_combined_prevalence}
\alias{compute_combined_prevalence}
\title{Compute combined prevalence of acute malnutrition}
\usage{
compute_pps_based_combined_prevalence(
  df,
  .wt = NULL,
  .edema = NULL,
  .summary_by
)

compute_combined_prevalence(df, .wt = NULL, .edema = NULL, .summary_by = NULL)
}
\arguments{
\item{df}{A data frame object returned by \code{\link[=process_muac_data]{process_muac_data()}} and \code{\link[=process_wfhz_data]{process_wfhz_data()}}.
The process_***_data function will have to used both to prepare the input data to be used
in the \code{compute_combined_prevalence()}. The order of which comes first does not matter,
however, since the muac data processor transforms MUAC values into centimeters, those
need to be put back into millimeter. This can be achieved my using \code{\link[=recode_muac]{recode_muac()}} inside
\code{\link[dplyr:mutate]{dplyr::mutate()}} or \code{\link[base:transform]{base::transform()}} (see example number 3 below).}

\item{.wt}{A numeric vector containing survey weights. If set to NULL (default)
the function will assume self weights, like in ENA for SMART, if otherwise given, the
weighted analysis will be computed.}

\item{.edema}{A character vector containing child's status on edema with "n" for no
edema, "y" = yes edema. Should you data be coded differently, re-code it to aforementioned
codes.}

\item{.summary_by}{A character vector containing data on the geographical areas where
the data was collected. If you are working on a single survey data, set
.summary_by = NULL (default). If this argument is not used, the function will error.}
}
\value{
A tibble. The length vary depending on .summary_by. If set to NULL, a tibble of
1 x 16 is returned, otherwise, a tibble of n rows (depending on the number of geographical
areas in the data set) x 17.
}
\description{
\code{compute_combined_prevalence()} is handy function to compute the combined prevalence of
acute malnutrition using the WFHZ and the absolute values of MUAC and edema for case
definition. Under the hood, before prevalence computations begin, it first evaluates the
status of WFHZ, MFAZ's standard deviation and age ratio test, as documented in
\code{\link[=compute_wfhz_prevalence]{compute_wfhz_prevalence()}} and \code{\link[=compute_muac_prevalence]{compute_muac_prevalence()}}. Then, it decides on the
appropriate analysis approach to employ depending on the outcome of the aforementioned
checks: (i) if either WFHZ, MFAZ standard deviation as well as age ratio test are not
simultaneously problematic, a complex sample-based prevalence analysis (for a two-stage
PPS cluster sampling) is computed; (ii) all other possibilities will involve either one
of the z-scores or the age ratio test being problematic, thus NA (for Not Applicable)
get thrown to output table.

A concept of "combined flags" is introduced here. This consists on creating a new vector
(cflags) of the same length as the input vectors (wfhz_flags and mfaz_flags) and assesses
if any element of either input vector is a flag (1), then that element is labelled as
flag (1) in the "cflags" vector, otherwise is not flag (0). This ensures that all
flagged observations in the WFHZ data and in MFAZ data are excluded for the combined
prevalence analysis.
}
\examples{

## When .summary_by and .wt are set to NULL ----
p <- compute_combined_prevalence(
df = anthro.02,
.wt = NULL,
.edema = edema,
.summary_by = NULL
)

print(p)

## When .wt is not set to NULL ----
x <- compute_combined_prevalence(
df = anthro.02,
.wt = "wtfactor",
.edema = edema,
.summary_by = NULL
)

print(x)

## When working on data frame with multiple survey areas ----
s <- anthro.03 |>
process_age(age = age) |>
process_muac_data(
sex = sex,
muac = muac,
age = "age",
.recode_sex = TRUE,
.recode_muac = TRUE,
unit = "cm"
) |>
dplyr::mutate(muac = recode_muac(muac, unit = "mm")) |>
process_wfhz_data(
sex = sex,
weight = weight,
height = height,
.recode_sex = TRUE) |>
compute_combined_prevalence(
.edema = edema,
.summary_by = district
)

print(s)

}
