% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_month_to_days}
\alias{compute_month_to_days}
\title{Recode age variable from months to days}
\usage{
compute_month_to_days(x)
}
\arguments{
\item{x}{A numeric vector containing values of age in months.}
}
\value{
A numeric vector with values corresponding to age in days
}
\description{
Recode age variable from months to days
}
