% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_muac.R
\name{compute_muac_prevalence}
\alias{compute_muac_prevalence}
\title{Compute acute malnutrition prevalence based on MUAC (the absolute values)}
\usage{
compute_muac_prevalence(df, .wt = NULL, .edema = NULL, .summary_by = NULL)
}
\arguments{
\item{df}{A data frame object returned by \code{\link[=process_muac_data]{process_muac_data()}}.}

\item{.wt}{A numeric vector containing survey weights. If set to NULL (default) and
the function will assume self weighted, like in ENA for SMART, otherwise if given, the
weighted analysis will be computed with weighted population returned.}

\item{.edema}{A character vector containing child's status on edema with "n" for no
edema, "y" = yes edema. Should you data be coded differently, re-code it to aforementioned
codes.}

\item{.summary_by}{A character vector containing data on the geographical areas where
the data was collected. If you are working on a single survey data, set
.summary_by = NULL (default). If this argument is not used, the function will error.}
}
\value{
A tibble. The length vary depending on .summary_by. If set to NULL, a tibble of
1 x 16 is returned, otherwise, a tibble of n rows (depending on the number of geographical
areas in the data set) x 17.
}
\description{
\code{compute_muac_prevalence()} is a handy function designed to dynamically compute acute
malnutrition's prevalence using the absolute values of MUAC, but with use of MFAZ for
quality checks before advancing to prevalence computations. Under the hood, the function
first checks the status of MFAZ's standard deviation (SD) after removing flags, and
the status of age ratio among children aged 6:23 vs 24:59 months. Then it decides on the
appropriate prevalence analysis approach to follow: (i) if SD & age ratio are both not
problematic, a complex sample-based prevalence analysis (for a two-stage  PPS
cluster sampling) is computed; (ii) if MFAZ's SD is not problematic, but age ratio test
is, the CDC/SMART MUAC tool weighting approach is use to compute the prevalence; (iii)
lastly, if MFAZ's SD is problematic even if age ratio test is not, no prevalenec
analysis is computed and NA's (of Not Applicable) are thrown.
The function also super handy to work on large data sets with multiple survey areas. For
this, the aforementioned conditionals are checked for each survey areas in a summarised
data frame and prevalence analysis computed according to each row's scenario.
}
\examples{
## When .summary.by = NULL ----

x <- compute_muac_prevalence(
df = anthro.04,
.wt = NULL,
.edema = edema,
.summary_by = NULL
)

print(x)

## When .summary_by is not set to NULL ----

p <- compute_muac_prevalence(
df = anthro.04,
.wt = NULL,
.edema = edema,
.summary_by = province
)

print(x)


}
