% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_wfhz.R
\name{compute_pps_based_wfhz_prevalence}
\alias{compute_pps_based_wfhz_prevalence}
\title{Compute a weight-for-height, MUAC-for-age z-score, and MUAC based prevalence estimates
of data collected from a two-stage cluster survey sample design, with the
first stage sampling done with Probability Proportional to the size of population}
\usage{
compute_pps_based_wfhz_prevalence(df, .wt = NULL, .edema = NULL, .summary_by)
}
\arguments{
\item{df}{A data frame object returned by \code{\link[=process_wfhz_data]{process_wfhz_data()}} or \code{\link[=process_muac_data]{process_muac_data()}}.
this will contain the wrangled vectors that are read inside the function.}

\item{.wt}{A numeric vector containing survey weights. If set to NULL (default) and
the function will assume self weighted, like in ENA for SMART, otherwise if given, the
weighted analysis will be computed with weighted population returned.}

\item{.edema}{A character vector containing child's status on edema with "n" for no
edema, "y" = yes edema. Should you data be coded differently, re-code it to aforementioned
codes.}

\item{.summary_by}{A character vector containing data on the geographical areas where
the data was collected. This is to group the survey design object into different
geographical areas in the data and allow for summaries to be computed for each of them.}
}
\description{
Create a survey design object using the \code{\link[srvyr:as_survey_design]{srvyr::as_survey_design()}} and then calculate
the survey means as well the sum of positive cases.
}
