% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_wfhz.R
\name{compute_wfhz_prevalence}
\alias{compute_wfhz_prevalence}
\title{Compute acute malnutrition prevalence based on weight-for-height z-scores (WFHZ)}
\usage{
compute_wfhz_prevalence(df, .wt = NULL, .edema = NULL, .summary_by = NULL)
}
\arguments{
\item{df}{A data frame object returned by \code{\link[=process_wfhz_data]{process_wfhz_data()}}.}

\item{.wt}{A numeric vector containing survey weights. If set to NULL (default) and
the function will assume self weighted, like in ENA for SMART, otherwise if given, the
weighted analysis will be computed with weighted population returned.}

\item{.edema}{A character vector containing child's status on edema with "n" for no
edema, "y" = yes edema. Should you data be coded differently, re-code it to aforementioned
codes.}

\item{.summary_by}{A character vector containing data on the geographical areas where
the data was collected. If you are working on a single survey data, set
.summary_by = NULL (default).}
}
\value{
A tibble. The length vary depending on .summary_by. If set to NULL, a tibble of
1 x 16 is returned, otherwise, a tibble of n rowns (depending on the number of geographical
areas in the dataset) x 17.
}
\description{
\code{compute_wfhz_prevalence()} is a handy function designed to dynamically compute acute
malnutrition's prevalence using WFHZ. Under the hood, it first checks the status of
WFHZ's standard deviation (SD) after removing flags, and then it decides on the
appropriate prevalence analysis approach to follow: if SD is anything between excellent
and acceptable, a complex sample-based prevalence analysis (for a two-stage  PPS
cluster sampling) is computed, otherwise, a re-calculated prevalence using PROBIT method
with a sample mean and a SD = 1 is computed. On the former analysis approach, the function
was also designed to work around survey weights. Furthermore, \code{compute_wfhz_prevalence()}
is quite handy for using on a large data sets with multiple survey areas. For this,
the function checks the WFHZ's SD for each survey area and then computes each survey
area's prevalence according to the status of SD.
}
\examples{

## When .summary_by = NULL ----
anthro.03 |>
process_wfhz_data(
sex = sex,
weight = weight,
height = height,
.recode_sex = TRUE
) |>
compute_wfhz_prevalence(
.wt = NULL,
.edema = edema,
.summary_by = NULL
)

## When .summary_by is not set to NULL ----
anthro.03 |>
process_wfhz_data(
sex = sex,
weight = weight,
height = height,
.recode_sex = TRUE
) |>
compute_wfhz_prevalence(
.wt = NULL,
.edema = edema,
.summary_by = district
)

## When a weighted analysis is needed ----
anthro.02 |>
compute_wfhz_prevalence(
.wt = "wtfactor",
.edema = edema,
.summary_by = province
)

}
