% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_age_in_months}
\alias{get_age_in_months}
\title{Get age in months from birth-date and the data when data was collected.}
\usage{
get_age_in_months(DoS = NULL, DoB = NULL, age)
}
\arguments{
\item{DoS, DoB}{Vectors containing dates. \code{DoS} refers to the day, month
and year when the data was collected; while \code{DoB} refers to the date when the
child was born (birth-date).}

\item{age}{A numeric vector containing already given age in months, usually an
integer in the input data as it is estimated using local event calendars.}
}
\value{
A vector of name \code{age} storing age in months, a mix of double and
integer and \code{NA} for missing value if any of the processed age in months is
< 6 or > 59.99 months.
}
\description{
\code{get_age_in_months()} works inside \code{\link[dplyr:mutate]{dplyr::mutate()}} or \code{\link[base:transform]{base::transform()}}
}
