% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_scorers.R
\name{get_quality_score}
\alias{get_quality_score}
\title{Get the overall WHZ or MFAZ's quality score}
\usage{
get_quality_score(df, method = c("mfaz", "whz"))
}
\arguments{
\item{df}{A data frame containing the scores. If you wish the get the overall
quality score for MFAZ, the input data frame must have seven (7) required
columns containing test classification of flagged data, sex ratio, age ratio,
standard deviation, skewness, kurtosis, crude MUAC's digit preference.
Alternatively, if you wish to get the quality score of WHZ, then the input
data frame must have the exact same columns in the plausibility report of the
ENA for SMART software.}

\item{method}{The method you wish to get the overall quality score for.
A choice between "mfaz" and "whz". If you wish to know the overall survey
score of your WHZ data, set \code{method = whz}, otherwise set \code{method = mfaz} for
MFAZ. If by mistake a different input choice is given, an error will be
thrown with a message guiding how to go about.}
}
\value{
A vector (named \code{"quality_score"}) with the overall quality scores.
}
\description{
\code{get_quality_score()} provides the overall quality score of either WHZ or MFAZ,
by adding up the scores across each test criteria. This is an input to
\code{\link[=classify_overall_quality]{classify_overall_quality()}}.
}
\examples{
# example code
## Creat a `df` object ----

df <- data.frame(
flagged_class = "Excellent",
age_ratio_class = "Good",
sex_ratio_class = "Problematic",
dps_class = "Excellent",
sd_class = "Excellent",
skew_class = "Good",
kurt_class = "Acceptable"
)

## Apply get_quality_score()
get_quality_score(df, method = "mfaz")

# You can also choose to chain the functions with a pipe operator ----
df |>
get_quality_score(method = "mfaz")

}
