% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mfaz.01}
\alias{mfaz.01}
\title{A MUAC screening data from an anonymized setting}
\format{
A tibble with 661 rows and 4 columns.\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{sex} \tab Sex, "m" = boys, "f" = girls \cr
   \emph{months} \tab calculated age in months with two decimal places \cr
   \emph{edema} \tab Edema, "n" = no, "y" = yes \cr
   \emph{muac} \tab Mid-upper arm circumference (mm) \cr
}
}
\usage{
mfaz.01
}
\description{
A MUAC screening data from an anonymized setting
}
\examples{
mfaz.01

}
\keyword{datasets}
