% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{muac_data}
\alias{muac_data}
\title{A MUAC screening data from an anonymized setting}
\format{
A data frame with 4 columns and 661 rows\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{sex} \tab Sex, "m" = boys, "f" = girls \cr
   \emph{months} \tab calculated age in months with two decimal places \cr
   \emph{edema} \tab Edema, "n" = no, "y" = yes \cr
   \emph{muac} \tab Mid-upper arm circumference (mm) \cr
}
}
\usage{
muac_data
}
\description{
A MUAC screening data from an anonymized setting
}
\examples{
muac_scre

}
\keyword{datasets}
