% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize_zscore}
\alias{normalize_zscore}
\title{Normalize Zscores to follow a normal distribution}
\usage{
normalize_zscore(x)
}
\arguments{
\item{x}{A numeric (double) vector storing zscores (with a least 3 decimal
places)}
}
\value{
A numeric vector of the same length as the input vector
}
\description{
According to the \href{https://smartmethodology.org/}{SMART Methodology.},
when standard deviation of weight-for-height zscores is greater than 1.2,
the acute malnutrition prevalence is re-calculated using a standard deviation
of 1.0.
}
\details{
\code{normalize_zscore()} helps to achieve this end by normalizing the zscores. It
subtracts each element in a vector of zscores with the mean of the same vector
and then devides by the standard deviation of the same vector.
}
