% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_checkers.R
\name{check_plausibility_mfaz}
\alias{check_plausibility_mfaz}
\alias{check_plausibility_wfhz}
\alias{check_plausibility_muac}
\title{Plausibility checkers: MUAC-for-age z-scores, Weight-for-Height z-scores and
MUAC}
\usage{
check_plausibility_mfaz(df, sex, muac, age, flags, area)

check_plausibility_wfhz(df, sex, age, weight, height, flags, area)

check_plausibility_muac(df, flags, sex, muac)
}
\arguments{
\item{df}{A data frame object returned by \code{\link[=process_muac_data]{process_muac_data()}} for
\code{check_plausibility_mfaz()} and \code{check_plausibility_muac()} and returned by
\code{\link[=process_wfhz_data]{process_wfhz_data()}} for \code{check_plausibility_wfhz()}.}

\item{sex}{A vector telling whether a given child is a boy or girl.}

\item{muac}{A vector containing MUAC measurements.}

\item{age}{A vector containing children's age in months.}

\item{flags}{A character vector telling whether or not an observation is an
outlier.}

\item{area}{A vector with values on where was the data collected. If you are
analyzing a data set with just one area, provide it anyway to
\code{check_plausibility_mfaz()} or \code{check_plausibility_wfhz()}}

\item{weight}{A vector containing weight measurements in kilograms.}

\item{height}{A vector containing height measurements in centimeters.}
}
\value{
A summarized data frame containing quality checks statistics and
respective classification.
}
\description{
\code{check_plausibility_mfaz()}, \code{check_plausibility_wfhz()} and
\code{check_plausibility_muac()} lets you know the quality of your data, based on
the statistics around MUAC-for-age zscores, weight-for-height z-scores and on
crude MUAC, respectively. Note that \code{check_plausibility_wfhz()} is all about
WHZ only. If you wish to know about MUAC checks consider using either
\code{check_plausibility_mfaz()} or \code{check_plausibility_muac()}
}
\examples{

## Check Plausibility: MFAZ ----

anthro.01 |>
process_age(
svdate = "dos",
birdate = "dob",
age = age
) |>
process_muac_data(
sex = sex,
age = "age",
muac = muac,
.recode_sex = TRUE,
.recode_muac = TRUE,
unit = "cm"
) |>
check_plausibility_mfaz(
flags = flag_mfaz,
sex = sex,
muac = muac,
age = age,
area = area
)

## Check Plausibility: WFHZ ----

anthro.01 |>
process_age(
svdate = "dos",
birdate = "dob",
age = age
) |>
process_wfhz_data(
sex = sex,
weight = weight,
height = height,
.recode_sex = TRUE
) |>
check_plausibility_wfhz(
sex = sex,
age = age,
weight = weight,
height = height,
flags = flag_wfhz,
area = area
)

## Check Plausibility: MUAC ----

anthro.01 |>
process_muac_data(
sex = sex,
muac = muac,
age = NULL,
.recode_sex = TRUE,
.recode_muac = FALSE,
unit = "none"
) |>
check_plausibility_muac(
flags = flag_muac,
sex = sex,
muac = muac
)

}
