% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prev_data}
\alias{prev_data}
\title{District level representative survey data with zscores calculated}
\format{
A data frame with 10 columns and 7740 rows\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{area} \tab Location where the survey took place \cr
   \emph{cluster} \tab Primary sampling unit \cr
   \emph{sex} \tab Sex, "m" = boys, "f" = girls \cr
   \emph{age} \tab calculated age in months with two decimal places \cr
   \emph{edema} \tab Edema, "n" = no, "y" = yes \cr
   \emph{muac} \tab Mid-upper arm circumference (mm) \cr
   \emph{wfhz} \tab Weight-for-height z-scores with 3 decimal places \cr
   \emph{flag_wfhz} \tab Flagged observations. 1=flagged, 0=not flagged \cr
   \emph{mfaz} \tab MUAC-for-age z-scores with 3 decimal places \cr
   \emph{flag_mfaz} \tab Flagged observations: 1=flagged, 0=not flagged \cr
}
}
\usage{
prev_data
}
\description{
District level representative survey data with zscores calculated
}
\examples{
prev_data

}
\keyword{datasets}
