% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processors.R
\name{process_muac_data}
\alias{process_muac_data}
\title{Process MUAC data a get it ready for analyses}
\usage{
process_muac_data(
  df,
  sex,
  muac,
  age = NULL,
  .recode_sex = TRUE,
  .recode_muac = TRUE,
  unit = c("cm", "mm", "none")
)
}
\arguments{
\item{df}{The input data frame with variables sex, age and MUAC.}

\item{sex}{A vector storing values about whether the child is a boy or a girl.
The variable name must be named sex, otherwise it will not work.}

\item{muac}{A vector storing crude MUAC values.}

\item{age}{A vector storing values about child's age in months. The variable
name must be named age, otherwise it will not work. For instance, if given as
following: age = months it will not work.}

\item{.recode_sex}{Logical. It asks whether you should recode your sex variable
to the required shape to use in \code{process_muac_data()}. The default values for
sex are 1 for boys and 2 for girls. Setting \code{.recode_sex = TRUE} works on "m"
and "f" values. If your vector is in any different shape, you should put it in
"m" and "f" or right away to 1 or 2. If you are using data exported from ENA for
SMART software, then you should leave \code{.recode_sex} at its default: \code{TRUE}.}

\item{.recode_muac}{Logical. Choose between \code{TRUE} or \code{FALSE} if you wish or
not to recode the MUAC variable into the required format to work on.}

\item{unit}{A choice of the units to which you wish to convert your MUAC
variable into.}
}
\value{
A data frame of the same length as the input data, but with a
different width as explained:When \code{age} is available in the input data and
supplied, \code{process_muac_data} will return as output a data frame with two
new variables \code{mfaz} and \code{flags}. \code{mfaz} stores MUAC-for-age z-score (MFAZ)
values and \code{flags} tells you whether a given z-score is an outlier or not.
This job is done by \code{\link[=flag_outliers]{flag_outliers()}}. If age is not available in the input
data, therefore not possible to supply in this function, \code{process_muac_data}
will only return \code{flags}. This will refer to flags based on crude MUAC.
}
\description{
\code{process_muac_data()} gets your input data ready for downstream MUAC related
analysis.
}
\examples{

## Have a sample data ----

df <- data.frame(
 survey_date = as.Date(c(
 "2023-01-01", "2023-01-01", "2023-01-01", "2023-01-01", "2023-01-01")),
 birthdate = as.Date(c(
 "2019-01-01", NA, "2018-03-20", "2019-11-05", "2021-04-25")),
 age = c(NA, 36, NA, NA, NA),
 sex = c("m", "f", "m", "m", "f"),
 muac = c(110, 130, 300, 123, 125)
 )

 ## Apply function ----
 df |>
 process_age(
 svdate = "survey_date",
 birdate = "birthdate",
 age = age
 ) |>
 process_muac_data(
 sex = sex,
 age = "age",
 muac = muac,
 .recode_sex = TRUE,
 .recode_muac = TRUE,
 unit = "cm"
 )

}
