% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processors.R
\name{process_whz_data}
\alias{process_whz_data}
\title{Process Weight-for-Height data get it ready for analyses}
\usage{
process_whz_data(df, sex, weight, height, .recode_sex = TRUE)
}
\arguments{
\item{df}{The input data frame with variables sex, age and MUAC.}

\item{sex}{A vector storing values about whether the child is a boy or a girl.}

\item{weight, height}{Vectors storing weight values in kilograms and height
values in centimeters, respectively.}

\item{.recode_sex}{Logical. It asks whether you should recode your sex variable
to the required shape to use in \code{process_whz_data()}. The default values for
sex are 1 = boys and 2 = girls. Setting \code{.recode_sex = TRUE} works on "m"
and "f" values. If your vector is in any different shape, you should put it in
"m" and "f" or right away to 1 or 2. If you are using data exported from ENA for
SMART software, then you should leave \code{.recode_sex} at its default: \code{TRUE}.}
}
\value{
A data frame of the same length as the input data, but with a different
width: two new variables \code{wfhz} and \code{flags}. \code{wfhz} stores weight-for-height
z-score values with three decimal places. \code{flags} tells you whether a given
z-score is an outlier or not. This job is done by \code{\link[=flag_outliers]{flag_outliers()}}.
}
\description{
\code{process_whz_data()} gets your input data ready for downstream WHZ related
analysis.
}
\examples{
## Have a sample data ----
anthro.01 |>
process_whz_data(
sex = sex,
weight = weight,
height = height,
.recode_sex = TRUE
)

}
