% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{recode_month_to_day}
\alias{recode_month_to_day}
\title{Recode age variable from months to days}
\usage{
recode_month_to_day(x)
}
\arguments{
\item{x}{A numeric vector containing values of age in months.}
}
\value{
A numeric vector with values corresponding to age in days
}
\description{
Recode age variable from months to days
}
