% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processors.R
\name{recode_muac}
\alias{recode_muac}
\title{Recode crude MUAC variable into either centimeters or millimeters}
\usage{
recode_muac(muac, unit = c("cm", "mm"))
}
\arguments{
\item{muac}{A numeric vector storing values for MUAC that can be in centimeters
or in millimeters.}

\item{unit}{A choice of the units to which you wish to convert your MUAC
variable into.}
}
\value{
A transformed vector into the unit you wish to have.
}
\description{
Sometimes, a vector containing MUAC values may be in centimeters or in
millimeters. You may want to get in the right format to use with
\link[zscorer:addWGSR]{zscorer::addWGSR} or \code{\link[nipnTK:digitPreference]{nipnTK::digitPreference()}}. \code{recode_muac()} helps you
getting the vector in the right format for the job! It works inside works
inside \code{\link[dplyr:mutate]{dplyr::mutate()}} or \code{\link[base:transform]{base::transform()}}.
}
\examples{
# Have an input data with muac in mm ----
muac <- seq(90, 250, by = 4)

# Apply recode ----
recode_muac(muac, unit = "cm")

# Have an input data with muac in mm ----
muac <- seq(9.0, 25.0, by = 0.2)

# Apply recode ----
recode_muac(muac, unit = "mm")

}
