% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_flags}
\alias{remove_flags}
\title{Remove detected outliers}
\usage{
remove_flags(x)
}
\arguments{
\item{x}{A numeric vector containing zscore or crude MUAC values}
}
\value{
A vector of same size, with flagged data replaced by \code{NA}s.
}
\description{
\code{remove_flags()} removes flags detected by \code{\link[=flag_outliers]{flag_outliers()}}. It helps you
compute your statistics when flags needs to be removed, such as in standard
deviation.
}
