% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentable_df_outputs.R
\name{return_pretty_ouput_crude_muac}
\alias{return_pretty_ouput_crude_muac}
\title{Get a beatified and presentable output table}
\usage{
return_pretty_ouput_crude_muac(df)
}
\arguments{
\item{df}{An output data frame returned by \code{\link[=check_plausibility_crude_muac]{check_plausibility_crude_muac()}}.}
}
\value{
An output data frame of the same size as the input, but with values
formatted, columns renamed, and ready to share.
}
\description{
You may want to share the plausibility report in a table. You usually care for
a well formatted and beauty table, with values rounded, scientific notations
converted into conventional notations, etc. \code{return_pretty_ouput_crude_muac()} does
that for you so you already.
}
\examples{

quality <- data.01 |>
process_muac_data(
sex = "sex",
muac = "muac",
.recode_sex = TRUE,
.recode_muac = FALSE,
unit = "none"
) |>
check_plausibility_crude_muac() |>
return_pretty_ouput_crude_muac()

## View output ----
quality

}
