% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentable_df_outputs.R
\name{return_pretty_ouput_whz}
\alias{return_pretty_ouput_whz}
\title{Get a beatified and presentable output table}
\usage{
return_pretty_ouput_whz(df)
}
\arguments{
\item{df}{An output data frame returned by \code{\link[=check_plausibility_whz]{check_plausibility_whz()}}.}
}
\value{
An output data frame of the same size as the input, but with values
formatted, columns renamed, and ready to share.
}
\description{
You may want to share the plausibility report in a table. You usually care for
a well formatted and beauty table, with values rounded, scientific notations
converted into conventional notations, etc. \code{return_pretty_ouput_whz()} does
that for you so you already.
}
\examples{

quality <- data.01 |>
process_age() |>
process_whz_data(
sex = "sex",
wgt = "weight",
hgt = "height",
.recode_sex = TRUE
) |>
check_plausibility_whz(
flags = "flags",
sex = "sex",
age = "age",
wgt = "weight",
hgt = "hgt",
area = area
) |>
return_pretty_ouput_whz()

}
