% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_helpers.R
\name{tell_muac_analysis_strategy}
\alias{tell_muac_analysis_strategy}
\title{A helper function to tell how to go about MUAC prevalence analysis based on
on the output of age ratio and standard deviation test results}
\usage{
tell_muac_analysis_strategy(age_ratio_class, sd_class)
}
\arguments{
\item{age_ratio_class, sd_class}{Character vectors storing age ratio's p-values
and standard deviation's classification, respectively.}
}
\value{
A character vector of the same length containing the indication of
what to do for the MUAC prevalence analysis: "weighted", "unweighted" and
"missing". If "weighted", the CDC weighting approach is applied to correct for
age bias. If "unweighted" a normal complex sample analysis is applied, and for
the latter, NA are thrown.
}
\description{
A helper function to tell how to go about MUAC prevalence analysis based on
on the output of age ratio and standard deviation test results
}
