% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_classifiers.R
\name{classify_overall_quality}
\alias{classify_overall_quality}
\title{Get the overall data quality classification}
\usage{
classify_overall_quality(df)
}
\arguments{
\item{df}{A data frame containing a vector with the quality scores generated by
\code{\link[=compute_quality_score]{compute_quality_score()}}.}
}
\value{
A character vector of the same length, but a different width as the
input \code{df} is returned with a new column called \code{quality_class}.
}
\description{
\code{classify_overall_quality()} helps you in knowing the overall status of your
data quality. It classifies the overall score generated by
\code{\link[=compute_quality_score]{compute_quality_score()}} into four categories, as it is done in the
\href{https://smartmethodology.org/}{SMART Methodology},
namely: "Excellent", "Good", "Acceptable" and "Problematic". Beware that
the overall classification should be used as an indication to further
scrutinize of data before taking the decision to validate or invalidate the
results.
}
