% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_helpers_complex_sample.R
\name{define_wasting}
\alias{define_wasting}
\title{Add acute malnutrition case-definitions to the data frame}
\usage{
define_wasting(
  df,
  zscore = NULL,
  muac = NULL,
  edema = NULL,
  base = c("wfhz", "muac", "combined")
)
}
\arguments{
\item{df}{The data frame object containing the vectors with zscores, muac and
edema.}

\item{zscore}{The vector storing zscores values with 3 decimal places.}

\item{muac}{An integer vector containing MUAC measurements in mm.}

\item{edema}{A character vector of "y" = Yes, "n" = No bilateral edema.
Default is NULL.}

\item{base}{A choice of options to which you case definition should be based on.}
}
\value{
A data frame with three vectors added to the input data frame: "gam",
"sam" and "mam". If base = "combined" the vector names change to "cgam",
"csam" and "cmam" for combined global, severe and moderate acute malnutrition
respectively.
}
\description{
Use \code{define_wasting()} to add the case-definitions in your input data frame.
}
\examples{
# MUAC-based case-definition ----
x <- prev_data |>
define_wasting(
muac = muac,
edema = edema,
base = "muac"
)
head(x)

# Weight-for-height based case-definition ----
x <- prev_data |>
define_wasting(
zscore = wfhz,
edema = edema,
base = "wfhz"
)
head(x)

# Combined case-definition ----
x <- prev_data |>
define_wasting(
zscore = wfhz,
muac = muac,
edema = edema,
base = "combined"
)
head(x)

}
