% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_helpers_complex_sample.R
\name{normalize_zscore}
\alias{normalize_zscore}
\title{Normalize zscores to follow a normal distribution}
\usage{
normalize_zscore(x)
}
\arguments{
\item{x}{A numeric (double) vector storing zscores (with a least 3 decimal
places)}
}
\value{
A numeric vector of the same length as the input vector
}
\description{
According to the \href{https://smartmethodology.org/}{SMART Methodology.},
when standard deviation of weight-for-height zscores is problematic,
the acute malnutrition prevalence is re-calculated after normalizing the
zscores to get a standard deviation of 1.0.

\code{normalize_zscore()} helps to achieve this end by normalizing the zscores. It
subtracts each element in a vector of zscores with the observed mean of the
same vector and then divides by the observed standard deviation of the same
vector.
}
