% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processors.R
\name{remove_flags}
\alias{remove_flags}
\title{Remove detected outliers}
\usage{
remove_flags(x, unit = c("zscore", "crude"))
}
\arguments{
\item{x}{A numeric vector containing zscore or crude MUAC values.}

\item{unit}{A choice of the units to which you wish remove flags on. variable into.}
}
\value{
A vector of same size, with flagged data replaced by \code{NA}s.
}
\description{
\code{remove_flags()} removes flags detected by \code{\link[=flag_outliers]{flag_outliers()}}. It helps you
compute your statistics when flags needs to be removed, such as in standard
deviation.
}
